Lecture_2/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic-settings
pwdlib
PyJWT